added more questions

+++ src/com/company/_12_StringMatching/RabinKarp.java
@@ -14,11 +14,13 @@ public class RabinKarp {
 
         for (int i = 0; i <= str.length() - pattern.length(); i++) {
             if(patHash == strHash &&
-            checkEqual(str, pattern, i,i+pattern.length()-1,0,pattern.length()-1)){
+            checkEqual(str, pattern, i,i+pattern.length()-1,
+                    0,pattern.length()-1)){
                 System.out.print(i + " ");
             }
             if(i < str.length()-pattern.length()){
-                strHash = recalculateHash(str,i,i+pattern.length(),strHash,pattern.length());
+                strHash = recalculateHash(str,i,i+pattern.length(),
+                        strHash,pattern.length());
             }
         }
     }
@@ -26,19 +28,21 @@ public class RabinKarp {
     public static long createHashValue(String str, int n){
         long result = 0;
         for (int i = 0; i < n; i++) {
-            result += (long)(str.charAt(i)*(long)Math.pow(prime, i));
+            result += str.charAt(i)*(long)Math.pow(prime, i);
         }
         return result;
     }
 
-    public static long recalculateHash(String str, int oldInd, int newInd, long oldHash, int patlength){
+    public static long recalculateHash(String str, int oldInd,
+                                       int newInd, long oldHash, int patlength){
         long newHash = oldHash - str.charAt(oldInd);
         newHash /= prime;
-        newHash += (long)(str.charAt(newInd)*(long)(Math.pow(prime,patlength-1)));
+        newHash += str.charAt(newInd)*(long)(Math.pow(prime,patlength-1));
         return newHash;
     }
 
-    public static boolean checkEqual(String s1, String s2, int start1, int end1, int start2, int end2){
+    public static boolean checkEqual(String s1, String s2, int start1,
+                                     int end1, int start2, int end2){
         if(end1 - start1 != end2 - start2){
             return false;
         }
@@ -52,3 +56,5 @@ public class RabinKarp {
         return true;
     }
 }
+
+
